+++ app/app/__init__.py
@@ -1,0 +1,4 @@
+from .celery import app as celery_app
+
+
+all =['celery_app']
