+++ app/app/asgi.py
@@ -12,8 +12,8 @@ import os
 from django.core.asgi import get_asgi_application
 from channels.routing import ProtocolTypeRouter,URLRouter
 from channels.auth import AuthMiddlewareStack
-from scrape.routing import ws_urlpatterns
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'core.settings')
+from core.routing import ws_urlpatterns
+os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'app.settings')
