+++ app/app/celery.py
@@ -2,7 +2,7 @@ import os
 from celery import Celery
 
 
-os.environ.setdefault('DJANGO_SETTINGS_MODULE','core.settings')
+os.environ.setdefault('DJANGO_SETTINGS_MODULE','app.settings')
 
 
 app =Celery('core')
