--- app/core/__init__.py
@@ -1,4 +1,0 @@
-from .celery import app as celery_app
-
-
-all =['celery_app']
