+++ app/core/consumers.py
@@ -2,16 +2,16 @@ from unittest import result
 from channels.generic.websocket import AsyncJsonWebsocketConsumer
 import json
 
-class ScrapeConsumer(AsyncJsonWebsocketConsumer):
+class CoreConsumer(AsyncJsonWebsocketConsumer):
 
 
     async def connect(self):
-      await self.channel_layer.group_add('scrape', self.channel_name)
+      await self.channel_layer.group_add('core', self.channel_name)
       await self.accept()
 
 
     async def disconnect(self,code):
-         await self.channel_layer.group_discard('scrape', self.channel_name)
+         await self.channel_layer.group_discard('core', self.channel_name)
 
     async def scrape_result (self,event):
         results =event['text']
