+++ app/core/routing.py
@@ -1,0 +1,9 @@
+from django.urls import path
+from .consumers import CoreConsumer
+
+
+ws_urlpatterns =[
+
+    path('ws/core/', CoreConsumer.as_asgi())
+
+]
