--- app/scrape/routing.py
@@ -1,9 +1,0 @@
-from django.urls import path
-from .consumers import ScrapeConsumer
-
-
-ws_urlpatterns =[
-
-    path('ws/scrape/', ScrapeConsumer.as_asgi())
-
-]
